chore: move heatmap logic out of map component
